+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Value.hs
@@ -62,7 +62,8 @@ module Plutarch.LedgerApi.Value (
 
 import Plutarch.Builtin (PDataNewtype (PDataNewtype))
 import Plutarch.DataRepr (DerivePConstantViaData (DerivePConstantViaData))
-import Plutarch.Internal.Builtin (pand', pbuiltinIfThenElse)
+import Plutarch.Internal.Builtin (pand', pbuiltinIfThenElse, punsafeDowncast)
+import Plutarch.Internal.Term (punsafeCoerce)
 import Plutarch.LedgerApi.AssocMap qualified as AssocMap
 import Plutarch.LedgerApi.Utils (Mret)
 import Plutarch.Lift (
@@ -74,7 +75,6 @@ import Plutarch.Lift (
 import Plutarch.List qualified as List
 import Plutarch.Prelude hiding (psingleton)
 import Plutarch.TryFrom (PTryFrom (PTryFromExcess, ptryFrom'))
-import Plutarch.Unsafe (punsafeCoerce, punsafeDowncast)
 import PlutusLedgerApi.V1.Value qualified as PlutusValue
 import PlutusLedgerApi.V3 qualified as Plutus
 import PlutusTx.Prelude qualified as PlutusTx