+++ CHANGELOG.md
@@ -54,6 +54,8 @@
   to deprecation
 * `Plutarch.String` as it no longer contained anything useful
 * `Plutarch.Bool`, as it no longer contained anything useful
+* `Plutarch.Num`, `Plutarch.Integer`, as they were unified into a new module
+* `Plutarch.Unsafe`, as it no longer contained anything useful
 
 ### Fixed